aws cleaner: remove cluster at start
Initial implementation assumption was that cluster should be removed
at the end of the cleanup proccess. This allows to restart cleanup
proccess from scratch in case of dependency cleanup issues.

But for some EKS setups we might have ENIs attached directly to a cluster.
That blocks VPC deletion while ENIs exist. The ENIs are deleted
automatically after cluster deletion.
So, I moved cluster drop to the beginning of the proccess.